feat: add TodosCompletedSection component and integrate into TodosWidget
- Introduced TodosCompletedSection component to manage completed todos.
- Refactored TodosWidget to utilize TodosCompletedSection for displaying completed todos.
- Enhanced error handling with ErrorBar component in various widgets.
- Implemented FilterToolbar in PullRequestsPage, QuickLinksPage, RepositoriesPage, TodosPage, and WorkflowsPage for improved filtering and searching.
- Refactored WorkflowExecutionModal and WorkflowInputModal to use Modal component for consistent UI.
- Moved getScanIssueLabel function to a new utility file for better organization.
