+++ src/web/src/pages/todos/TodosPage.tsx
@@ -3,6 +3,7 @@ import { useQuery } from "@tanstack/react-query";
 import { todosApi } from "../../api/todos";
 import { useTodos } from "../../hooks/useTodos";
 import { TodosWidget } from "../dashboard/widgets/TodosWidget";
+import { FilterToolbar } from "../../components/FilterToolbar";
 import type { TodoItem } from "../../types";
 import "./TodosPage.css";
 
@@ -30,36 +31,25 @@ export default function TodosPage() {
     })
     .filter((t) => !search || t.title.toLowerCase().includes(search.toLowerCase()));
 
+  const todoFilters = [
+    { value: "all", label: `All (${todos.length})` },
+    { value: "active", label: `Active${activeCount > 0 ? ` (${activeCount})` : ""}` },
+    { value: "completed", label: `Completed${completedCount > 0 ? ` (${completedCount})` : ""}` },
+  ];
+
   return (
     <div className="todos-page">
       <div className="todos-page-card">
-        <div className="todos-page-toolbar">
-          <div className="todos-page-toolbar-left">
-            <input
-              className="todos-page-search"
-              type="search"
-              placeholder="Search todos…"
-              value={search}
-              onChange={(e) => setSearch(e.target.value)}
-            />
-            <div className="todos-page-filters">
-              {(["all", "active", "completed"] as Filter[]).map((f) => (
-                <button
-                  key={f}
-                  className={`todos-page-filter-btn${filter === f ? " todos-page-filter-btn--active" : ""}`}
-                  onClick={() => setFilter(f)}
-                >
-                  {f === "all" && `All (${todos.length})`}
-                  {f === "active" && `Active${activeCount > 0 ? ` (${activeCount})` : ""}`}
-                  {f === "completed" && `Completed${completedCount > 0 ? ` (${completedCount})` : ""}`}
-                </button>
-              ))}
-            </div>
-          </div>
-          {filtered.length !== todos.length && (
-            <span className="todos-page-count">{filtered.length} shown</span>
-          )}
-        </div>
+        <FilterToolbar
+          search={search}
+          onSearchChange={setSearch}
+          filter={filter}
+          onFilterChange={(f) => setFilter(f as Filter)}
+          filters={todoFilters}
+          searchPlaceholder="Search todos…"
+          shownCount={filtered.length}
+          totalCount={todos.length}
+        />
       </div>
 
       {isLoading ? (
