+++ src/web/src/pages/workflows/WorkflowsPage.tsx
@@ -1,25 +1,19 @@
 import { useState } from "react";
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import { workflowsApi } from "../../api/workflows";
 import { WorkflowsWidget } from "../dashboard/widgets/WorkflowsWidget";
 import { WorkflowInputModal } from "../dashboard/widgets/WorkflowInputModal";
 import { WorkflowExecutionModal } from "../dashboard/widgets/WorkflowExecutionModal";
-import type { WorkflowDefinition, WorkflowExecution, RunWorkflowRequest } from "../../types";
+import { FilterToolbar } from "../../components/FilterToolbar";
+import { useWorkflowModals } from "../../hooks/useWorkflowModals";
+import type { WorkflowDefinition, WorkflowExecution } from "../../types";
 import "./WorkflowsPage.css";
 
 type Filter = "all" | "idle" | "running" | "succeeded" | "failed";
 
 export default function WorkflowsPage() {
-  const queryClient = useQueryClient();
   const [search, setSearch] = useState("");
   const [filter, setFilter] = useState<Filter>("all");
-  const [workflowRunError, setWorkflowRunError] = useState<string | null>(null);
-  const [workflowInputModal, setWorkflowInputModal] = useState<WorkflowDefinition | null>(null);
-  const [runningWorkflowId, setRunningWorkflowId] = useState<string | null>(null);
-  const [workflowModal, setWorkflowModal] = useState<{
-    workflow: WorkflowDefinition;
-    executionId: string | null;
-  } | null>(null);
 
   const { data: workflows = [], isLoading } = useQuery<WorkflowDefinition[]>({
     queryKey: ["workflows"],
@@ -32,31 +26,19 @@ export default function WorkflowsPage() {
     refetchInterval: 5000,
   });
 
-  const runWorkflow = useMutation({
-    mutationFn: ({ workflowId, request }: { workflowId: string; request: RunWorkflowRequest }) => {
-      setRunningWorkflowId(workflowId);
-      return workflowsApi.run(workflowId, request);
-    },
-    onSuccess: (execution, variables) => {
-      setWorkflowRunError(null);
-      queryClient.invalidateQueries({ queryKey: ["workflow-executions"] });
-      const workflow = workflows.find((w) => w.id === variables.workflowId);
-      if (workflow) {
-        setWorkflowModal({ workflow, executionId: execution.id });
-      }
-    },
-    onError: (err) => setWorkflowRunError(err.message),
-    onSettled: () => setRunningWorkflowId(null),
+  const {
+    workflowRunError,
+    setWorkflowRunError,
+    workflowInputModal,
+    setWorkflowInputModal,
+    runningWorkflowId,
+    workflowModal,
+    setWorkflowModal,
+    runWorkflowWithInputs,
+  } = useWorkflowModals({
+    workflows,
+    confirmMessage: (workflow) => `Run workflow "${workflow.name}"?`,
   });
-
-  function runWorkflowWithInputs(workflow: WorkflowDefinition, inputs: Record<string, string>) {
-    const confirmed =
-      !workflow.requiresConfirmation ||
-      window.confirm(`Run workflow "${workflow.name}"?`);
-    if (!confirmed) return;
-    setWorkflowRunError(null);
-    runWorkflow.mutate({ workflowId: workflow.id, request: { inputs, confirmed } });
-  }
 
   function getStatus(workflow: WorkflowDefinition): string {
     return workflowExecutions.find((e) => e.workflowId === workflow.id)?.status ?? "idle";
@@ -76,36 +58,27 @@ export default function WorkflowsPage() {
         w.description?.toLowerCase().includes(search.toLowerCase()),
     );
 
+  const workflowFilters = [
+    { value: "all", label: `All (${workflows.length})` },
+    ...( ["idle", "running", "succeeded", "failed"] as const).map((s) => ({
+      value: s,
+      label: `${s.charAt(0).toUpperCase() + s.slice(1)}${statusCounts[s] > 0 ? ` (${statusCounts[s]})` : ""}`,
+    })),
+  ];
+
   return (
     <div className="workflows-page">
       <div className="workflows-page-card">
-        <div className="workflows-page-toolbar">
-          <div className="workflows-page-toolbar-left">
-            <input
-              className="workflows-page-search"
-              type="search"
-              placeholder="Search by name or description…"
-              value={search}
-              onChange={(e) => setSearch(e.target.value)}
-            />
-            <div className="workflows-page-filters">
-              {(["all", "idle", "running", "succeeded", "failed"] as Filter[]).map((f) => (
-                <button
-                  key={f}
-                  className={`workflows-page-filter-btn${filter === f ? " workflows-page-filter-btn--active" : ""}`}
-                  onClick={() => setFilter(f)}
-                >
-                  {f === "all"
-                    ? `All (${workflows.length})`
-                    : `${f.charAt(0).toUpperCase() + f.slice(1)}${statusCounts[f] > 0 ? ` (${statusCounts[f]})` : ""}`}
-                </button>
-              ))}
-            </div>
-          </div>
-          {filtered.length !== workflows.length && (
-            <span className="workflows-page-count">{filtered.length} shown</span>
-          )}
-        </div>
+        <FilterToolbar
+          search={search}
+          onSearchChange={setSearch}
+          filter={filter}
+          onFilterChange={(f) => setFilter(f as Filter)}
+          filters={workflowFilters}
+          searchPlaceholder="Search by name or description…"
+          shownCount={filtered.length}
+          totalCount={workflows.length}
+        />
       </div>
 
       {isLoading ? (
